Fix Terminate scope bug
Fixed a problem with the logic in the if else statements from lines 94 to 101 prior to this commit:
In case the rule of self is 'bidirectional', the if statement on line 94-95 evaluates to true. The code following this statement will update the scope of self only if obj is downstream of self in the
sequence. But for bidirectional, the scope also has to be updated if obj is upstream of self. This was meant to happen in the next elif statement.

However: the conditional of line 98-99 is an elif statement. This leads to that this code is skipped if the self.rule is bidirectional, since the if block of line 94-95 evaluated to true.

So if the bidirectional terminate modifier (obj) is upstream of the tagobject (self) it should modify, that tagobject's (self) scope won't be modified.